updated read data structure to have double fragment lengths

+++ src/read.hpp
@@ -1,12 +1,13 @@
-
+#include <iostream>
+#include <vector>
 
 class Read{
 
 	public:	
-	std::vector<unsigned int> fragments;
+	std::vector<double> fragments;
 	std::string name;
 	std::string enzyme;
 	std::string something;	
 
-	void printRead();
+	void printRead();	
 };
